🔧 WOW-40 fix : error 수정
